backend/requirements.txt
yfinance>=0.2.33
pandas>=2.2.0
numpy>=1.26.0
ta>=0.11.0
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
google-generativeai>=0.8.5
apscheduler>=3.10.0
